requirements.txt
okta==2.1.0
python-gitlab==2.10.1
requests==2.25.0
